Add custom expectation for Carbon instances and level cap validation
A custom expectation for Carbon instances named 'toBeCarbon' was added in the 'tests/Pest.php' file to improve date format tests. It was inspired by the similar expectation in the Pest Plugin Test Time.

In the 'src/Concerns/GiveExperience.php' file, a check was included to ensure that the points awarded do not exceed the experience requirement for the last level. This validation helps keep the game balanced and prevents unreasonable levels of experience. The respective test was added in the 'tests/Concerns/GiveExperienceTest.php' file.

In the 'tests/Concerns/HasStreaksTest.php' file, all the date comparisons were updated to use the new 'toBeCarbon' expectation. This makes these tests cleaner and easier to understand.