Refactor experience and level-up systems
Updated the level-up system to appropriately fire events for each level gained and not just the final level. Removed the direct update of the user's level within the levelUp method and instead implemented a function to calculate and update to the highest possible level based on a user's current points. This ensures that levelling up and firing of events occur for each level achieved, rather than skipping to the final level. Additionally, methods within the GiveExperience.php file were rearranged for better readability and maintenance.